Refactor App.tsx
