+++ template/index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './src/app/App';
+import App from './src/App';
 import {name as appName} from './src/app.json';
 
 AppRegistry.registerComponent(appName, () => App);
