+++ template/src/__tests__/App-test.tsx
@@ -3,7 +3,7 @@
  */
 
 import 'react-native';
-import App from '../app/App';
+import App from '../App';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
